After all, it's the Hello, World
In the past summer vacation, I made the promise to myself that I should
at least be a beginner in Python, for it's for "beginners". Well, it
really confuses me, of whether continue using C and learn C++, or at
least have some knowledge of Python.

The answer is......Both.